+++ packages/extension-base/ctp-extension-azure-adapters.js
