+++ packages/extension-base/ctp-extension-gcf-adapters.js
