+++ packages/extension-base/ctp-extension-gcf-adapters.spec.js
@@ -1,9 +1,8 @@
-const adapter = require('./ctp-extension-azure-adapter');
+const adapters = require('./ctp-extension-gcf-adapters');
 
 const createResponse = () => ({ json: jest.fn() });
 const createContext = response => ({
   log: jest.fn(),
-  done: jest.fn(),
   status: jest.fn(() => response),
 });
 const createRequest = () => ({
@@ -22,7 +21,7 @@ describe('when creating', () => {
     response = createResponse();
     context = createContext(response);
 
-    adapter.createExtensionAdapter(fn)(context, request);
+    adapters.createExtensionAdapter(fn)(context, request);
   });
 
   it('should invoke the `fn`', () => {
@@ -47,22 +46,12 @@ describe('when creating', () => {
       fn.mock.calls[0][1].accept(actions);
     });
 
-    it('should add a `res` to the context', () => {
-      expect(context.res).toBeDefined();
+    it('should set status to `200`', () => {
+      expect(context.status).toHaveBeenCalledWith(200);
     });
 
-    it('should add a `status` `200` to the `res`', () => {
-      expect(context.res.status).toBe(200);
-    });
-
-    it('should add a `body` with `actions` to the `res`', () => {
-      expect(context.res.body).toEqual({
-        actions,
-      });
-    });
-
-    it('should invoke `done` on the context', () => {
-      expect(context.done).toHaveBeenCalled();
+    it('should pass actions to response', () => {
+      expect(response.json).toHaveBeenCalledWith({ actions });
     });
   });
 
@@ -73,22 +62,12 @@ describe('when creating', () => {
       fn.mock.calls[0][1].reject(errors);
     });
 
-    it('should add a `res` to the context', () => {
-      expect(context.res).toBeDefined();
+    it('should set status to `200`', () => {
+      expect(context.status).toHaveBeenCalledWith(400);
     });
 
-    it('should add a `status` `400` to the `res`', () => {
-      expect(context.res.status).toBe(400);
-    });
-
-    it('should add a `body` with `errors` to the `res`', () => {
-      expect(context.res.body).toEqual({
-        errors,
-      });
-    });
-
-    it('should invoke `done` on the context', () => {
-      expect(context.done).toHaveBeenCalled();
+    it('should pass errors to response', () => {
+      expect(response.json).toHaveBeenCalledWith({ errors });
     });
   });
 });
